[Model] hf model modification and inheritance change
