+++ src/lmflow/models/hf_decoder_model.py
@@ -44,18 +44,17 @@ from peft import (
     get_peft_model,
     prepare_model_for_kbit_training
 )
-from peft.utils.constants import TRANSFORMERS_MODELS_TO_LORA_TARGET_MODULES_MAPPING
 
 from lmflow.datasets.dataset import Dataset
+from lmflow.models.hf_model_mixin import HFModelMixin
 from lmflow.models.decoder_model import DecoderModel
 from lmflow.models.interfaces.tunable import Tunable
 from lmflow.utils.constants import (
     TEXT_ONLY_DATASET_DESCRIPTION,
     TEXT2TEXT_DATASET_DESCRIPTION,
     CONVERSATION_DATASET_DESCRIPTION,
-    LMFLOW_LORA_TARGET_MODULES_MAPPING
 )
-from lmflow.utils.conversation_template import ConversationTemplate, PRESET_TEMPLATES
+from lmflow.utils.conversation_template import PRESET_TEMPLATES
 from lmflow.tokenization.hf_decoder_model import (
     tokenize_function, 
     conversation_tokenize_function
@@ -64,11 +63,6 @@ from lmflow.tokenization.hf_decoder_model import (
 
 logger = logging.getLogger(__name__)
 
-
-LORA_TARGET_MODULES_MAPPING_MIXIN = {
-    k: TRANSFORMERS_MODELS_TO_LORA_TARGET_MODULES_MAPPING.get(k, LMFLOW_LORA_TARGET_MODULES_MAPPING.get(k)) 
-    for k in set(TRANSFORMERS_MODELS_TO_LORA_TARGET_MODULES_MAPPING) | set(LMFLOW_LORA_TARGET_MODULES_MAPPING)
-}
 
 MODELS_SUPPORT_FLASH_ATTENTION = [
     "LlamaForCausalLM",
@@ -99,7 +93,7 @@ except Exception as e:
     else:
         logger.warning(f'An error occurred when importing flash_attn, flash attention is disabled: {e}')
 
-class HFDecoderModel(DecoderModel, Tunable):
+class HFDecoderModel(DecoderModel, HFModelMixin, Tunable):
     r"""
     Initializes a HFDecoderModel instance.
 
@@ -138,266 +132,15 @@ class HFDecoderModel(DecoderModel, Tunable):
         :param tune_strategy: tuning strategy: normal, none, lora or adapter
         :param ds_config: deepspeed configuration for distributed training
         """
-
-        # See more about loading any type of standard or custom dataset (from
-        # files, python dict, pandas DataFrame, etc) at
-        # https://huggingface.co/docs/datasets/loading_datasets.html.
-
-        # Load pretrained model and tokenizer
-        #
-        # Distributed training: The .from_pretrained methods guarantee that
-        # only one local process can concurrently download model & vocab.
-
-        self.device = device
-        self.model_args = model_args
-        tokenizer_kwargs = {
-            "cache_dir": model_args.cache_dir,
-            "use_fast": model_args.use_fast_tokenizer,
-            "revision": model_args.model_revision,
-            "use_auth_token": True if model_args.use_auth_token else None,
-            "trust_remote_code": model_args.trust_remote_code,
-        }
-        
-        try:
-            if model_args.tokenizer_name:
-                tokenizer = AutoTokenizer.from_pretrained(model_args.tokenizer_name, **tokenizer_kwargs)
-            elif model_args.model_name_or_path:
-                tokenizer = AutoTokenizer.from_pretrained(model_args.model_name_or_path, **tokenizer_kwargs)
-            else:
-                raise ValueError(
-                    "You are instantiating a new tokenizer from scratch. This is"
-                    " not supported by this script. You can do it from another"
-                    " script, save it, and load it from here, using"
-                    " --tokenizer_name."
-                )
-
-        except RecursionError:
-            logger.warning("The tokenizer_config.json file doesn't set the special tokens. Using default values: <unk>, <s>, </s> for unknown token, bos token and eos token respectively.")
-            if model_args.tokenizer_name:
-                tokenizer = AutoTokenizer.from_pretrained(model_args.tokenizer_name, unk_token="<unk>",
-                                                    bos_token="<s>",
-                                                    eos_token="</s>",
-                                                    **tokenizer_kwargs)
-            elif model_args.model_name_or_path:
-                tokenizer = AutoTokenizer.from_pretrained(model_args.model_name_or_path, unk_token="<unk>",
-                                                    bos_token="<s>",
-                                                    eos_token="</s>",
-                                                    **tokenizer_kwargs)
-            else:
-                raise ValueError(
-                    "You are instantiating a new tokenizer from scratch. This is"
-                    " not supported by this script. You can do it from another"
-                    " script, save it, and load it from here, using"
-                    " --tokenizer_name."
-                )
-            
-        self.tokenizer = tokenizer  
-
-        torch_dtype = (
-            model_args.torch_dtype
-            if model_args.torch_dtype in ["auto", None]
-            else getattr(torch, model_args.torch_dtype)
+        super(HFModelMixin, self).__init__(
+            model_args=model_args,
+            do_train=True if tune_strategy == "normal" else False,
+            ds_config=ds_config,
+            device=device,
+            use_accelerator=use_accelerator,
+            *args,
+            **kwargs
         )
-        logger.debug(f"torch_dtype on init: {torch_dtype}")
-
-        config_kwargs = {
-            "cache_dir": model_args.cache_dir,
-            "revision": model_args.model_revision,
-            "use_auth_token": True if model_args.use_auth_token else None,
-            "trust_remote_code": model_args.trust_remote_code,
-        }
-        if model_args.config_name:
-            config = AutoConfig.from_pretrained(model_args.config_name, **config_kwargs)
-        elif model_args.model_name_or_path:
-            config = AutoConfig.from_pretrained(model_args.model_name_or_path, **config_kwargs)
-        else:
-            config = CONFIG_MAPPING[model_args.model_type]()
-            logger.warning("You are instantiating a new config instance from scratch.")
-            if model_args.config_overrides is not None:
-                logger.info(f"Overriding config: {model_args.config_overrides}")
-                config.update_from_string(model_args.config_overrides)
-                logger.info(f"New config: {config}")
-
-        #position interpolation
-        if model_args.do_rope_scaling:
-            if "LlamaForCausalLM" in config.architectures:
-                from lmflow.utils.position_interpolation.llama_rope_scaled_monkey_patch import (
-                        replace_llama_with_condense,
-                )
-                replace_llama_with_condense(model_args.rope_pi_ratio, model_args.rope_ntk_ratio)
-
-        if tune_strategy == 'normal':
-            if model_args.model_name_or_path:
-                compute_dtype = torch_dtype
-                device_map = "auto"
-                if os.environ.get('LOCAL_RANK') is not None:
-                    local_rank = int(os.environ.get('LOCAL_RANK','0'))
-                    device_map = {'': local_rank}
-
-                if model_args.use_qlora:
-                    model_args.use_lora = True
-                    quant_config = BitsAndBytesConfig(
-                        load_in_4bit=model_args.bits == 4,
-                        load_in_8bit=model_args.bits == 8,
-                        llm_int8_threshold=6.0,
-                        llm_int8_has_fp16_weight=False,
-                        bnb_4bit_compute_dtype=compute_dtype,
-                        bnb_4bit_use_double_quant=model_args.double_quant,
-                        bnb_4bit_quant_type=model_args.quant_type,
-                    )
-                try:
-                    model = AutoModelForCausalLM.from_pretrained(
-                        model_args.model_name_or_path,
-                        from_tf=bool(".ckpt" in model_args.model_name_or_path),
-                        quantization_config=quant_config if model_args.use_qlora else None,
-                        cache_dir=model_args.cache_dir,
-                        revision=model_args.model_revision,
-                        use_auth_token=True if model_args.use_auth_token else None,
-                        torch_dtype=torch_dtype,
-                        trust_remote_code = model_args.trust_remote_code,
-                        attn_implementation="flash_attention_2" if model_args.use_flash_attention else None,
-                    )
-                #for deepspeed zero3, we don't need to specify device_map
-                except:
-                    model = AutoModelForCausalLM.from_pretrained(
-                        model_args.model_name_or_path,
-                        from_tf=bool(".ckpt" in model_args.model_name_or_path),
-                        config=config,
-                        quantization_config=quant_config if model_args.use_qlora else None,
-                        cache_dir=model_args.cache_dir,
-                        revision=model_args.model_revision,
-                        use_auth_token=True if model_args.use_auth_token else None,
-                        torch_dtype=torch_dtype,
-                        trust_remote_code = model_args.trust_remote_code,
-                        attn_implementation="flash_attention_2" if model_args.use_flash_attention else None,
-                    )
-                if model_args.use_qlora:
-                    model.gradient_checkpointing_enable()
-                    model = prepare_model_for_kbit_training(model)
-            else:
-                model = AutoModelForCausalLM.from_config(config)
-                n_params = sum(dict((p.data_ptr(), p.numel()) for p in model.parameters()).values())
-                logger.info(f"Training new model from scratch - Total size={n_params/2**20:.2f}M params")
-            self.backend_model_full = model
-            if model_args.use_lora:
-                if model_args.lora_target_modules:
-                    lora_target_modules = model_args.lora_target_modules
-                else:
-                    model_config = getattr(model, "config", {"model_type": "custom"})
-                    if hasattr(model_config, "to_dict"):
-                        model_config = model_config.to_dict()                    
-                    lora_target_modules = LORA_TARGET_MODULES_MAPPING_MIXIN.get(model_config["model_type"], None)
-                peft_config = LoraConfig(
-                    task_type=TaskType.CAUSAL_LM,
-                    inference_mode=False,
-                    r=model_args.lora_r,
-                    lora_alpha=model_args.lora_alpha,
-                    lora_dropout=model_args.lora_dropout,
-                    target_modules=lora_target_modules,
-                )
-                model = get_peft_model(model, peft_config)
-                model.print_trainable_parameters()
-
-            # We resize the embeddings only when necessary to avoid index errors.
-            # If you are creating a model from scratch on a small vocab and want a
-            # smaller embedding size, remove this test.
-            with deepspeed.zero.GatheredParameters(model.get_input_embeddings().weight, modifier_rank=None):
-                weights = model.get_input_embeddings().weight
-                embedding_size = weights.shape[0]
-            if len(tokenizer) > embedding_size:
-                model.resize_token_embeddings(len(tokenizer))
-
-            self.config = config
-            self.backend_model = model
-            self.tune_strategy = tune_strategy
-
-        elif tune_strategy == 'none':
-            if use_accelerator:
-                peft_model_id = model_args.lora_model_path
-                self.backend_model = AutoModelForCausalLM.from_pretrained(
-                        model_args.model_name_or_path,
-                        config=config,
-                        device_map="auto",
-                        offload_folder="offload",
-                        offload_state_dict=True,
-                        torch_dtype=torch_dtype,
-                        load_in_8bit = model_args.use_int8,
-                        attn_implementation="flash_attention_2" if model_args.use_flash_attention else None,
-                    )
-                if peft_model_id is not None:
-                    self.backend_model = PeftModel.from_pretrained(
-                        self.backend_model, 
-                        peft_model_id,
-                    )
-                self.tokenizer.padding_side = "left"
-            else:
-                dschf = HfDeepSpeedConfig(ds_config)
-                peft_model_id = model_args.lora_model_path
-                # NOTE: Currently offload is not supported by llama
-                if config.model_type == "llama" and model_args.use_ram_optimized_load:
-                    logger.warning(
-                        "llama does not support RAM optimized load. Automatically"
-                        " use original load instead."
-                    )
-                    model_args.use_ram_optimized_load = False
-
-                if model_args.use_ram_optimized_load and peft_model_id is None:
-                    try:
-                        # RAM-optimized load
-                        self.backend_model = AutoModelForCausalLM.from_pretrained(
-                            model_args.model_name_or_path,
-                            config=config,
-                            device_map="auto",
-                            offload_folder="offload",
-                            offload_state_dict=True,
-                            torch_dtype=torch_dtype,
-                            attn_implementation="flash_attention_2" if model_args.use_flash_attention else None,
-                        )
-                    except:
-                        logger.warning(
-                            "Failed to use RAM optimized load. Automatically"
-                            " use original load instead."
-                        )
-                        # Normal load
-                        self.backend_model = AutoModelForCausalLM.from_pretrained(
-                            model_args.model_name_or_path,
-                            config=config,
-                            torch_dtype=torch_dtype,
-                            attn_implementation="flash_attention_2" if model_args.use_flash_attention else None,
-                        )
-                else:
-                    if peft_model_id is not None:
-                        logger.warning(
-                            "LoRA does not support RAM optimized load currently."
-                            " Automatically use original load instead."
-                        )
-                    self.backend_model = AutoModelForCausalLM.from_pretrained(
-                        model_args.model_name_or_path,
-                        config=config,
-                        torch_dtype=torch_dtype,
-                        attn_implementation="flash_attention_2" if model_args.use_flash_attention else None,
-                    )
-
-                self.backend_model_full = self.backend_model
-                if peft_model_id is not None:
-                    self.backend_model = PeftModel.from_pretrained(
-                        self.backend_model, peft_model_id
-                    )
-  
-                self.tokenizer.padding_side = "left" #necessary for llama, gpt2 and other decoder models
-                
-                if device == "gpu":
-                    deepspeed.init_distributed()
-                    self.ds_engine = deepspeed.initialize(model=self.backend_model, config_params=ds_config)[0]
-                    self.ds_engine.module.eval()
-
-        elif tune_strategy == 'adapter':
-            raise NotImplementedError('adapter tune strategy not implemented')
-
-        if self.tokenizer.eos_token_id is None:
-            self.tokenizer.eos_token_id = self.backend_model.config.eos_token_id
-        if self.tokenizer.pad_token_id is None:
-            self.tokenizer.pad_token_id = self.tokenizer.eos_token_id
 
 
     def tokenize(self, dataset, add_special_tokens=True, *args, **kwargs):
@@ -725,24 +468,3 @@ class HFDecoderModel(DecoderModel, Tunable):
             logger.warning(f"Save full model with dtype: {save_dtype}")
         else:
             self.get_backend_model().save_pretrained(dir)
-
-
-    def get_max_length(self):
-        """
-        Return max acceptable input length in terms of tokens.
-        """
-        return self.tokenizer.model_max_length
-
-
-    def get_tokenizer(self):
-        """
-        Return the tokenizer of the model.
-        """
-        return self.tokenizer
-
-
-    def get_backend_model(self):
-        """
-        Return the backend model.
-        """
-        return self.backend_model
